refactor(#3266): extract operator-assignment subsystem from assignment.ts god-file (#3182)
Verbatim, behaviour-preserving leaf move of the `x op= y` lowering (22 fns:
logical assignment &&=/||=/??= + compound assignment +=/-=/*=/&=/>>= and the
string/native-string += fast paths) out of the 7504-LOC assignment.ts god-file
into a new sibling module src/codegen/expressions/operator-assignment.ts.

- Pure move: no logic changes. Byte-identity IDENTICAL (prove-emit-identity
  39/39 across gc/standalone/wasi). tsc 0 errors.
- Coupling is one-directional: the new module imports back only
  compileExternSetFallback (already exported) from ./assignment.js; no forward
  edge, no import cycle. assignment.ts imports are unchanged (every one is still
  used by what stays).
- 3 external importers (binary-ops.ts, expressions.ts, unary-updates.ts) repoint
  their imports of the 4 public entry points to the new module.
- assignment.ts: 7504 -> 4642 LOC; new module 2922 LOC.

Relocation-shift ratchets (byte-identity proves total usage is conserved):
- loc-budget + coercion-sites: per-issue frontmatter allowances in the #3266
  issue file (change-scoped gates).
- oracle-ratchet: 2 preauthorized entries (15 getTypeAtLocation + 19 ctx.checker
  RELOCATED sites; assignment.ts drops 21->6 / 32->13, conserved), mirroring the
  #808 precedent.

tests/issue-3266.test.ts smoke-tests both moved families (#2093 probe gate).

✓
