+++ plan/issues/3266-split-assignment.md
@@ -1,7 +1,8 @@
 ---
 id: 3266
 title: "Split operator-assignment subsystem out of assignment.ts god-file"
-status: ready
+status: done
+completed: 2026-07-14
 sprint: current
 priority: high
 feasibility: medium
@@ -10,6 +11,13 @@ task_type: refactor
 subtask_of: 3182
 area: codegen
 assignee: ttraenkler/sendev-splitassign
+# Relocation-shift allowances (#3131 change-scoped gates read these from the PR's own issue file).
+# Byte-identity IDENTICAL (prove-emit-identity 39/39) proves total usage is conserved — these are
+# false-positive relocation shifts (counts moved out of assignment.ts into the new sibling module).
+loc-budget-allow:
+  - src/codegen/expressions/operator-assignment.ts
+coercion-sites-allow:
+  - src/codegen/expressions/operator-assignment.ts
 ---
 
 # 3266 — Split operator-assignment subsystem out of assignment.ts god-file
@@ -35,12 +43,14 @@ zero-coupling leaf that moves verbatim into a NEW sibling module:
     `compilePropertyCompoundAssignment`, `compilePropertyCompoundAssignmentExternref`,
     `emitToPropertyKeyOnce`, `compileElementCompoundAssignment`.
 
-Coupling is zero: the moved functions reference no same-file top-level symbol, and the
-remaining file (plain `=` assignment + destructuring) never calls any moved function. Only 4
-public entry points are consumed externally (`compileLogicalAssignment`,
-`compileCompoundAssignment`, `isCompoundAssignment`, `emitToPropertyKeyOnce`) by 3 files
-(`binary-ops.ts`, `expressions.ts`, `unary-updates.ts`), which repoint their imports to the new
-module.
+Coupling is near-zero and one-directional: the moved functions reference exactly ONE
+staying-behind symbol — `compileExternSetFallback` (already exported by assignment.ts) — so the
+new module imports it back from `./assignment.js`; there is NO forward edge (the remaining file —
+plain `=` assignment + destructuring — never calls any moved function). This keeps a clean
+one-way edge (operator-assignment.ts → assignment.ts) with no import cycle. Only 4 public entry
+points are consumed externally (`compileLogicalAssignment`, `compileCompoundAssignment`,
+`isCompoundAssignment`, `emitToPropertyKeyOnce`) by 3 files (`binary-ops.ts`, `expressions.ts`,
+`unary-updates.ts`), which repoint their imports to the new module.
 
 ## Acceptance
 
@@ -49,6 +59,19 @@ module.
 - All relocation-shift ratchets green (allowances below preauthorize the false-positive
   relocation shifts; byte-identity IDENTICAL proves total usage is conserved).
 
-## Relocation-shift allowances
+## Relocation-shift allowances (applied)
 
-<!-- Filled in after running the ratchets locally; each is a documented per-issue preauth. -->
+All ratchets run locally; byte-identity IDENTICAL (39/39) is the proof these are false-positive
+relocation shifts (total usage conserved, counts merely moved module home):
+
+- **LOC budget** (`check:loc-budget`) — new module is 2922 LOC (> 1500). Granted via the
+  `loc-budget-allow:` frontmatter above. (assignment.ts drops 7504 → 4642.)
+- **Oracle ratchet** (`check:oracle-ratchet`) — 15 `getTypeAtLocation` + 19 `ctx.checker`
+  RELOCATED sites. Granted via two `preauthorized` entries in
+  `scripts/oracle-ratchet-baseline.json` (the only mechanism this gate reads; mirrors the #808
+  precedent). assignment.ts's counts drop by the same amount: getTypeAtLocation 21 → 6,
+  ctx.checker 32 → 13 (perfectly conserved).
+- **Coercion sites** (`check:coercion-sites`) — 12 RELOCATED coercion-vocabulary sites
+  (`number_toString` ×4, `emitBoolToString` ×3, `__unbox_number` ×5). Granted via the
+  `coercion-sites-allow:` frontmatter above.
+- **Dead exports** (`check:dead-exports`) and **verdict-oracle bump** — pass unchanged (pure move).
